+++ src/Tests/Aggregations/Bucket/Children/ChildrenAggregationMapping.doc.cs
