+++ src/Tests/ClientConcepts/HighLevel/CovariantHits/CovariantSearchResults.doc.cs
