+++ src/Tests/CommonOptions/DateMath/DateMathExpressions.doc.cs
@@ -10,7 +10,7 @@ using Nest;
 using Tests.Framework;
 using static Tests.Framework.RoundTripper;
 
-namespace Tests.CommonOptions
+namespace Tests.CommonOptions.DateMath
 {
 	public class DateMathEpressions
 	{
