+++ src/Tests/Document/Single/Index/IndexUrl.cs
@@ -10,7 +10,7 @@ using static Tests.Framework.UrlTester;
 
 namespace Tests.Document.Single.Index
 {
-	public class IndexUrlTests
+	public class IndexUrl
 	{
 		[U] public async Task Urls()
 		{
