+++ src/Tests/Tests.csproj
@@ -93,12 +93,12 @@
     <Compile Include="ClientConcepts\ConnectionPooling\Sniffing\OnConnectionFailure.doc.cs" />
     <Compile Include="ClientConcepts\ConnectionPooling\Sniffing\OnStaleClusterState.doc.cs" />
     <Compile Include="ClientConcepts\ConnectionPooling\Sniffing\RoleDetection.doc.cs" />
-    <Compile Include="ClientConcepts\HighLevel\CovariantHits\CovariantSearchResults.cs" />
+    <Compile Include="ClientConcepts\HighLevel\CovariantHits\CovariantSearchResults.doc.cs" />
     <Compile Include="ClientConcepts\HighLevel\Inferrence\DocumentPaths\DocumentPaths.doc.cs" />
-    <Compile Include="ClientConcepts\HighLevel\Inferrence\Id\IdsInference.cs" />
-    <Compile Include="ClientConcepts\HighLevel\Inferrence\Indices\IndicesPaths.cs" />
-    <Compile Include="CodeStandards\Descriptors.cs" />
-    <Compile Include="Document\Single\Index\IndexUrlTests.cs" />
+    <Compile Include="ClientConcepts\HighLevel\Inferrence\Id\IdsInference.doc.cs" />
+    <Compile Include="ClientConcepts\HighLevel\Inferrence\Indices\IndicesPaths.doc.cs" />
+    <Compile Include="CodeStandards\Descriptors.doc.cs" />
+    <Compile Include="Document\Single\Index\IndexUrl.cs" />
     <Compile Include="Document\Single\Index\Indexing.Doc.cs" />
     <Compile Include="Framework\CrudBase.cs" />
     <Compile Include="Framework\MockResponses\SniffingResponse.cs" />
@@ -127,13 +127,13 @@
     <Compile Include="Framework\VirtualClustering\VirtualCluster.cs" />
     <Compile Include="Framework\VirtualClustering\VirtualClusterConnection.cs" />
     <Compile Include="Framework\VirtualClustering\VirtualizedCluster.cs" />
-    <Compile Include="ClientConcepts\HighLevel\Inferrence\FieldNames\FieldInference.cs" />
-    <Compile Include="ClientConcepts\LowLevel\Connecting.cs" />
+    <Compile Include="ClientConcepts\HighLevel\Inferrence\FieldNames\FieldInference.doc.cs" />
+    <Compile Include="ClientConcepts\LowLevel\Connecting.doc.cs" />
     <Compile Include="ClientConcepts\ConnectionPooling\BuildingBlocks\KeepingTrackOfNodes.doc.cs" />
     <Compile Include="ClientConcepts\ConnectionPooling\BuildingBlocks\DateTimeProviders.doc.cs" />
-    <Compile Include="ClientConcepts\LowLevel\PostData.cs" />
-    <Compile Include="CommonOptions\DateMath\DateMathExpressions.cs" />
-    <Compile Include="CommonOptions\TimeUnit\TimeUnits.cs" />
+    <Compile Include="ClientConcepts\LowLevel\PostData.doc.cs" />
+    <Compile Include="CommonOptions\DateMath\DateMathExpressions.doc.cs" />
+    <Compile Include="CommonOptions\TimeUnit\TimeUnits.doc.cs" />
     <Compile Include="Framework\MockData\CommitActivity.cs" />
     <Compile Include="Framework\MockData\Developer.cs" />
     <Compile Include="Framework\MockData\Gender.cs" />
